fix mapingg to long instead of integer

+++ mom-core/src/main/java/net/aydini/mom/util/mapper/StringToIntegerMapper.java
@@ -3,17 +3,17 @@ package net.aydini.mom.util.mapper;
 import net.aydini.mom.common.exception.MomBaseException;
 import net.aydini.mom.common.service.maper.Maper;
 
-public class StringToIntegerMapper implements Maper<String, Long> {
+public class StringToIntegerMapper implements Maper<String, Integer> {
 
 	@Override
-	public Long map(String input) {
+	public Integer map(String input) {
 
 		if (input == null || input.isEmpty())
 			return null;
 
 		try {
 
-			return Long.parseLong(input);
+			return Integer.parseInt(input);
 		} catch (NumberFormatException e) {
 
 			throw new MomBaseException(e.getMessage(), e);
